fixing resolution of line trace camera
This resolution change was needed because the
line trace camera's FOV was too wide that it could
see it's robot's motor.

As my research, looks like changing the resolution
doesn't change the size of the FOV, it tries to
manage to get the largest size with the same aspect
ratio within the maximum size of 4608x2592, which
was very interesting and new for me.

Which means, I've still got to optimize the size
when I get back to my club tomorrow. Good night
for today.